+++ src/iter/monitor.c
@@ -98,7 +98,7 @@ struct iter_monitor_s* iter_monitor_create(long N, const float* image_truth, voi
 }
 
 
-void monitor_iter6(struct monitor_iter6_s* monitor, long epoch, long batch, long num_batches, float objective, long NI, const float* x[NI], char* post_string)
+void monitor_iter6(struct monitor_iter6_s* monitor, int epoch, int batch, int num_batches, float objective, long NI, const float* x[NI], char* post_string)
 {
 	if ((NULL != monitor) && (NULL != monitor->fun))
 		monitor->fun(monitor, epoch, batch, num_batches, objective, NI, x, post_string);
